Greatly simplify media metadata loading
Use a fake DOM element that isn't actually added to the DOM and reduce 
typing overhead
