+++ communityvi-frontend/src/lib/components/medium_selector/MediumSelector.svelte
@@ -4,7 +4,7 @@
 	import type {MediumStateChanged} from '$lib/client/model';
 	import {onDestroy} from 'svelte';
 	import {formatMediumLength} from '$lib/components/medium_selector/helpers';
-	import MetadataLoader, {SelectedMedium} from '$lib/components/medium_selector/metadata_loader';
+	import {MediumMetadata} from '$lib/components/medium_selector/medium_metadata';
 	import RegisteredClient from '$lib/client/registered_client';
 
 	interface Properties {
@@ -17,7 +17,6 @@
 	let medium: Medium | undefined = $state();
 	let mediumIsOutdated = $derived(medium !== undefined && videoUrl === undefined);
 
-	let durationHelper: HTMLVideoElement;
 	let fileSelector: HTMLInputElement;
 
 	let unsubscribe: (() => void) = $state(() => {});
@@ -53,11 +52,9 @@
 			return;
 		}
 
-		const metadataLoader = new MetadataLoader(durationHelper);
-
-		let selectedMedium: SelectedMedium;
+		let selectedMedium: MediumMetadata;
 		try {
-			selectedMedium = await metadataLoader.selectedMediumFromFile(selectedFile);
+			selectedMedium = await MediumMetadata.fromFile(selectedFile);
 		} catch (error) {
 			console.error('Error while loading medium:', error);
 			notifications.reportError(error as Error);
@@ -100,9 +97,6 @@
 		fileSelector.value = '';
 	}
 </script>
-
-<!-- Hidden video element for parsing file metadata -->
-<video hidden={true} muted={true} bind:this={durationHelper}></video>
 
 <section id="medium-selection">
 	<span class="medium-title">{medium?.name ?? 'n/a'}</span>
